split out special functions on FactorInstancesProvider to be adopters of it, and changing the FactorInstancesProvider to have fields and use methods rather than functions.
