+++ src/factor_instances_provider/provider/factor_instances_provider_unit_tests.rs
@@ -4,82 +4,6 @@ use std::ops::Add;
 
 use crate::{factor_instances_provider::provider::test_sargon_os::SargonOS, prelude::*};
 
-type Sut = FactorInstancesProvider;
-
-#[should_panic]
-#[actix_rt::test]
-async fn mfa_panics_if_entities_empty() {
-    let fs = HDFactorSource::fs0();
-    let a = Account::sample_unsecurified();
-    let _ = Sut::for_account_mfa(
-        &mut FactorInstancesCache::default(),
-        MatrixOfFactorSources::new([], 1, [fs.clone()]),
-        Profile::new([fs], [&a], []),
-        IndexSet::new(), // <---- EMPTY => should_panic
-        Arc::new(TestDerivationInteractors::default()),
-    )
-    .await
-    .unwrap();
-}
-
-#[should_panic]
-#[actix_rt::test]
-async fn mfa_panics_if_entity_unknown() {
-    let fs = HDFactorSource::fs0();
-    let a = Account::sample_unsecurified();
-    let _ = Sut::for_account_mfa(
-        &mut FactorInstancesCache::default(),
-        MatrixOfFactorSources::new([], 1, [fs.clone()]),
-        Profile::new([fs], [&a], []),
-        IndexSet::just(Account::a1().entity_address()), // <---- unknown => should_panic
-        Arc::new(TestDerivationInteractors::default()),
-    )
-    .await
-    .unwrap();
-}
-
-#[should_panic]
-#[actix_rt::test]
-async fn mfa_panics_if_wrong_network() {
-    let fs = HDFactorSource::fs0();
-    let network = NetworkID::Mainnet;
-    let mainnet_account = Account::unsecurified_on_network(
-        "main",
-        network,
-        HierarchicalDeterministicFactorInstance::tx_on_network(
-            CAP26EntityKind::Account,
-            network,
-            HDPathComponent::unsecurified_hardening_base_index(0),
-            fs.factor_source_id(),
-        ),
-    );
-    let network = NetworkID::Stokenet;
-    let stokenet_account = Account::unsecurified_on_network(
-        "stoknet",
-        network,
-        HierarchicalDeterministicFactorInstance::tx_on_network(
-            CAP26EntityKind::Account,
-            network,
-            HDPathComponent::unsecurified_hardening_base_index(0),
-            fs.factor_source_id(),
-        ),
-    );
-    let profile = Profile::new([fs.clone()], [&mainnet_account, &stokenet_account], []);
-    assert_eq!(profile.networks.len(), 2);
-    let _ = Sut::for_account_mfa(
-        &mut FactorInstancesCache::default(),
-        MatrixOfFactorSources::new([], 1, [fs.clone()]),
-        profile,
-        IndexSet::from_iter([
-            mainnet_account.entity_address(),
-            stokenet_account.entity_address(),
-        ]), // <---- wrong network => should_panic
-        Arc::new(TestDerivationInteractors::default()),
-    )
-    .await
-    .unwrap();
-}
-
 #[actix_rt::test]
 async fn create_accounts_when_last_is_used_cache_is_fill_only_with_account_vecis_and_if_profile_is_used_a_new_account_is_created(
 ) {
@@ -148,390 +72,6 @@ async fn create_accounts_when_last_is_used_cache_is_fill_only_with_account_vecis
         !os.cache_snapshot()
             .is_full(NetworkID::Mainnet, bdfs.factor_source_id()),
         "just consumed one, so not full"
-    );
-}
-
-#[actix_rt::test]
-async fn cache_is_always_filled_persona_veci_then_after_all_used_we_start_over_at_zero_if_no_profile_is_used(
-) {
-    let network = NetworkID::Mainnet;
-    let bdfs = HDFactorSource::sample();
-    let mut cache = FactorInstancesCache::default();
-
-    let outcome = Sut::for_persona_veci(
-        &mut cache,
-        None,
-        bdfs.clone(),
-        network,
-        Arc::new(TestDerivationInteractors::default()),
-    )
-    .await
-    .unwrap();
-
-    assert_eq!(outcome.factor_source_id, bdfs.factor_source_id());
-
-    assert_eq!(outcome.debug_found_in_cache.len(), 0);
-
-    assert_eq!(
-        outcome.debug_was_cached.len(),
-        DerivationPreset::all().len() * CACHE_FILLING_QUANTITY
-    );
-
-    assert_eq!(
-        outcome.debug_was_derived.len(),
-        DerivationPreset::all().len() * CACHE_FILLING_QUANTITY + 1
-    );
-
-    let instances_used_directly = outcome.to_use_directly.factor_instances();
-    assert_eq!(instances_used_directly.len(), 1);
-    let instances_used_directly = instances_used_directly.first().unwrap();
-
-    assert_eq!(
-        instances_used_directly.derivation_entity_index(),
-        HDPathComponent::Hardened(HDPathComponentHardened::Unsecurified(
-            UnsecurifiedIndex::unsecurified_hardening_base_index(0)
-        ))
-    );
-
-    cache.assert_is_full(network, bdfs.factor_source_id());
-
-    let cached = cache
-        .peek_all_instances_of_factor_source(bdfs.factor_source_id())
-        .unwrap();
-
-    let persona_veci_paths = cached
-        .clone()
-        .get(&DerivationPreset::IdentityVeci.index_agnostic_path_on_network(network))
-        .unwrap()
-        .factor_instances()
-        .into_iter()
-        .map(|x| x.derivation_path())
-        .collect_vec();
-
-    assert_eq!(persona_veci_paths.len(), CACHE_FILLING_QUANTITY);
-
-    assert!(persona_veci_paths
-        .iter()
-        .all(|x| x.entity_kind == CAP26EntityKind::Identity
-            && x.network_id == network
-            && x.key_space() == KeySpace::Unsecurified
-            && x.key_kind == CAP26KeyKind::TransactionSigning));
-
-    let persona_veci_indices = persona_veci_paths
-        .into_iter()
-        .map(|x| x.index)
-        .collect_vec();
-
-    assert_eq!(
-        persona_veci_indices.first().unwrap().clone(),
-        HDPathComponent::unsecurified_hardening_base_index(1)
-    );
-
-    assert_eq!(
-        persona_veci_indices.last().unwrap().clone(),
-        HDPathComponent::unsecurified_hardening_base_index(30)
-    );
-}
-
-#[actix_rt::test]
-async fn cache_is_always_filled_account_veci_then_after_all_used_we_start_over_at_zero_if_no_profile_is_used(
-) {
-    let network = NetworkID::Mainnet;
-    let bdfs = HDFactorSource::sample();
-    let mut cache = FactorInstancesCache::default();
-
-    let outcome = Sut::for_account_veci(
-        &mut cache,
-        None,
-        bdfs.clone(),
-        network,
-        Arc::new(TestDerivationInteractors::default()),
-    )
-    .await
-    .unwrap();
-
-    assert_eq!(outcome.factor_source_id, bdfs.factor_source_id());
-
-    assert_eq!(outcome.debug_found_in_cache.len(), 0);
-
-    assert_eq!(
-        outcome.debug_was_cached.len(),
-        DerivationPreset::all().len() * CACHE_FILLING_QUANTITY
-    );
-
-    assert_eq!(
-        outcome.debug_was_derived.len(),
-        DerivationPreset::all().len() * CACHE_FILLING_QUANTITY + 1
-    );
-
-    let instances_used_directly = outcome.to_use_directly.factor_instances();
-    assert_eq!(instances_used_directly.len(), 1);
-    let instances_used_directly = instances_used_directly.first().unwrap();
-
-    assert_eq!(
-        instances_used_directly.derivation_entity_index(),
-        HDPathComponent::Hardened(HDPathComponentHardened::Unsecurified(
-            UnsecurifiedIndex::unsecurified_hardening_base_index(0)
-        ))
-    );
-
-    cache.assert_is_full(network, bdfs.factor_source_id());
-
-    let cached = cache
-        .peek_all_instances_of_factor_source(bdfs.factor_source_id())
-        .unwrap();
-
-    let account_veci_paths = cached
-        .clone()
-        .get(&DerivationPreset::AccountVeci.index_agnostic_path_on_network(network))
-        .unwrap()
-        .factor_instances()
-        .into_iter()
-        .map(|x| x.derivation_path())
-        .collect_vec();
-
-    assert_eq!(account_veci_paths.len(), CACHE_FILLING_QUANTITY);
-
-    assert!(account_veci_paths
-        .iter()
-        .all(|x| x.entity_kind == CAP26EntityKind::Account
-            && x.network_id == network
-            && x.key_space() == KeySpace::Unsecurified
-            && x.key_kind == CAP26KeyKind::TransactionSigning));
-
-    let account_veci_indices = account_veci_paths
-        .into_iter()
-        .map(|x| x.index)
-        .collect_vec();
-
-    assert_eq!(
-        account_veci_indices.first().unwrap().clone(),
-        HDPathComponent::unsecurified_hardening_base_index(1)
-    );
-
-    assert_eq!(
-        account_veci_indices.last().unwrap().clone(),
-        HDPathComponent::unsecurified_hardening_base_index(30)
-    );
-
-    let account_mfa_paths = cached
-        .clone()
-        .get(&DerivationPreset::AccountMfa.index_agnostic_path_on_network(network))
-        .unwrap()
-        .factor_instances()
-        .into_iter()
-        .map(|x| x.derivation_path())
-        .collect_vec();
-
-    assert!(account_mfa_paths
-        .iter()
-        .all(|x| x.entity_kind == CAP26EntityKind::Account
-            && x.network_id == network
-            && x.key_space() == KeySpace::Securified
-            && x.key_kind == CAP26KeyKind::TransactionSigning));
-
-    let account_mfa_indices = account_mfa_paths.into_iter().map(|x| x.index).collect_vec();
-
-    assert_eq!(
-        account_mfa_indices.first().unwrap().clone(),
-        HDPathComponent::securifying_base_index(0)
-    );
-
-    assert_eq!(
-        account_mfa_indices.last().unwrap().clone(),
-        HDPathComponent::securifying_base_index(29)
-    );
-
-    let identity_mfa_paths = cached
-        .clone()
-        .get(&DerivationPreset::IdentityMfa.index_agnostic_path_on_network(network))
-        .unwrap()
-        .factor_instances()
-        .into_iter()
-        .map(|x| x.derivation_path())
-        .collect_vec();
-
-    assert!(identity_mfa_paths
-        .iter()
-        .all(|x| x.entity_kind == CAP26EntityKind::Identity
-            && x.network_id == network
-            && x.key_space() == KeySpace::Securified
-            && x.key_kind == CAP26KeyKind::TransactionSigning));
-
-    let identity_mfa_indices = identity_mfa_paths
-        .into_iter()
-        .map(|x| x.index)
-        .collect_vec();
-
-    assert_eq!(
-        identity_mfa_indices.first().unwrap().clone(),
-        HDPathComponent::securifying_base_index(0)
-    );
-
-    assert_eq!(
-        identity_mfa_indices.last().unwrap().clone(),
-        HDPathComponent::securifying_base_index(29)
-    );
-
-    let identity_veci_paths = cached
-        .clone()
-        .get(&DerivationPreset::IdentityVeci.index_agnostic_path_on_network(network))
-        .unwrap()
-        .factor_instances()
-        .into_iter()
-        .map(|x| x.derivation_path())
-        .collect_vec();
-
-    assert!(identity_veci_paths
-        .iter()
-        .all(|x| x.entity_kind == CAP26EntityKind::Identity
-            && x.network_id == network
-            && x.key_space() == KeySpace::Unsecurified
-            && x.key_kind == CAP26KeyKind::TransactionSigning));
-
-    let identity_veci_indices = identity_veci_paths
-        .into_iter()
-        .map(|x| x.index)
-        .collect_vec();
-
-    assert_eq!(
-        identity_veci_indices.first().unwrap().clone(),
-        HDPathComponent::unsecurified_hardening_base_index(0)
-    );
-
-    assert_eq!(
-        identity_veci_indices.last().unwrap().clone(),
-        HDPathComponent::unsecurified_hardening_base_index(29)
-    );
-
-    // lets create another account (same network, same factor source)
-
-    let outcome = Sut::for_account_veci(
-        &mut cache,
-        None,
-        bdfs.clone(),
-        network,
-        Arc::new(TestDerivationInteractors::default()),
-    )
-    .await
-    .unwrap();
-
-    assert_eq!(outcome.factor_source_id, bdfs.factor_source_id());
-    assert_eq!(outcome.debug_found_in_cache.len(), 1); // This time we found in cache
-    assert_eq!(outcome.debug_was_cached.len(), 0);
-    assert_eq!(outcome.debug_was_derived.len(), 0);
-
-    let instances_used_directly = outcome.to_use_directly.factor_instances();
-    assert_eq!(instances_used_directly.len(), 1);
-    let instances_used_directly = instances_used_directly.first().unwrap();
-
-    assert_eq!(
-        instances_used_directly.derivation_entity_index(),
-        HDPathComponent::Hardened(HDPathComponentHardened::Unsecurified(
-            UnsecurifiedIndex::unsecurified_hardening_base_index(1) // Next one!
-        ))
-    );
-
-    assert!(!cache.is_full(network, bdfs.factor_source_id())); // not full anymore, since we just used a veci
-
-    let cached = cache
-        .peek_all_instances_of_factor_source(bdfs.factor_source_id())
-        .unwrap();
-
-    let account_veci_paths = cached
-        .clone()
-        .get(&DerivationPreset::AccountVeci.index_agnostic_path_on_network(network))
-        .unwrap()
-        .factor_instances()
-        .into_iter()
-        .map(|x| x.derivation_path())
-        .collect_vec();
-
-    assert_eq!(account_veci_paths.len(), CACHE_FILLING_QUANTITY - 1);
-
-    assert!(account_veci_paths
-        .iter()
-        .all(|x| x.entity_kind == CAP26EntityKind::Account
-            && x.network_id == network
-            && x.key_space() == KeySpace::Unsecurified
-            && x.key_kind == CAP26KeyKind::TransactionSigning));
-
-    let account_veci_indices = account_veci_paths
-        .into_iter()
-        .map(|x| x.index)
-        .collect_vec();
-
-    assert_eq!(
-        account_veci_indices.first().unwrap().clone(),
-        HDPathComponent::unsecurified_hardening_base_index(2) // first is not `1` anymore
-    );
-
-    assert_eq!(
-        account_veci_indices.last().unwrap().clone(),
-        HDPathComponent::unsecurified_hardening_base_index(30)
-    );
-
-    // create 29 more accounts, then we should be able to crate one more which should ONLY derive
-    // more instances for ACCOUNT VECI, and not Identity Veci, Identity MFA and Account MFA, since that is
-    // not needed.
-    for _ in 0..29 {
-        let outcome = Sut::for_account_veci(
-            &mut cache,
-            None,
-            bdfs.clone(),
-            network,
-            Arc::new(TestDerivationInteractors::default()),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(outcome.factor_source_id, bdfs.factor_source_id());
-
-        assert_eq!(outcome.debug_found_in_cache.len(), 1);
-        assert_eq!(outcome.debug_was_cached.len(), 0);
-        assert_eq!(outcome.debug_was_derived.len(), 0);
-    }
-
-    let cached = cache
-        .peek_all_instances_of_factor_source(bdfs.factor_source_id())
-        .unwrap();
-
-    assert!(
-        cached
-            .get(&DerivationPreset::AccountVeci.index_agnostic_path_on_network(network))
-            .is_none(),
-        "should have used the last instance..."
-    );
-
-    // Great, now lets create one more account, and this time we should derive more instances for
-    // it. We should derive 31 instances, 30 for account veci to cache and 1 to use directly.
-    // we should NOT derive more instances for Identity Veci, Identity MFA and Account MFA, since
-    // that cache is already full.
-    let outcome = Sut::for_account_veci(
-        &mut cache,
-        None,
-        bdfs.clone(),
-        network,
-        Arc::new(TestDerivationInteractors::default()),
-    )
-    .await
-    .unwrap();
-
-    assert_eq!(outcome.factor_source_id, bdfs.factor_source_id());
-
-    assert_eq!(outcome.debug_found_in_cache.len(), 0);
-    assert_eq!(outcome.debug_was_cached.len(), CACHE_FILLING_QUANTITY); // ONLY 30, not 120...
-    assert_eq!(outcome.debug_was_derived.len(), CACHE_FILLING_QUANTITY + 1);
-
-    let instances_used_directly = outcome.to_use_directly.factor_instances();
-    assert_eq!(instances_used_directly.len(), 1);
-    let instances_used_directly = instances_used_directly.first().unwrap();
-
-    assert_eq!(
-        instances_used_directly.derivation_entity_index(),
-        HDPathComponent::Hardened(HDPathComponentHardened::Unsecurified(
-            UnsecurifiedIndex::unsecurified_hardening_base_index(0) // IMPORTANT! Index 0 is used again! Why?! Well because are not using a Profile here, and we are not eagerly filling cache just before we are using the last index.
-        ))
     );
 }
 
@@ -758,47 +298,6 @@ async fn adding_accounts_different_networks_different_factor_sources() {
             .map(|a| a.entity_address())
             .collect::<HashSet<AccountAddress>>()
     );
-}
-
-#[actix_rt::test]
-async fn securify_accounts_and_personas_with_override_factor() {
-    // this is mostly a soundness test for the two functions `for_persona_mfa` and `for_account_mfa`
-    // using `os` because I'm lazy. We might in fact remove `for_persona_mfa` and `for_account_mfa`
-    // and only use the `for_entity_mfa` function... but we have these to get code coverage.
-    let (mut os, bdfs) = SargonOS::with_bdfs().await;
-
-    let (batman, stats) = os.new_mainnet_persona_with_bdfs("Batman").await.unwrap();
-    assert!(stats.debug_was_derived.is_empty());
-
-    let (alice, stats) = os.new_mainnet_account_with_bdfs("alice").await.unwrap();
-    assert!(stats.debug_was_derived.is_empty());
-
-    let shield_0 = MatrixOfFactorSources::new([], 0, [bdfs.clone()]);
-    let mut cache = os.cache_snapshot();
-    let interactors = Arc::new(TestDerivationInteractors::default());
-    let outcome = FactorInstancesProvider::for_account_mfa(
-        &mut cache,
-        shield_0.clone(),
-        os.profile_snapshot(),
-        IndexSet::just(alice.entity_address()),
-        interactors.clone(),
-    )
-    .await
-    .unwrap();
-    let outcome = outcome.per_factor.get(&bdfs.factor_source_id()).unwrap();
-    assert_eq!(outcome.to_use_directly.len(), 1);
-
-    let outcome = FactorInstancesProvider::for_persona_mfa(
-        &mut cache,
-        shield_0.clone(),
-        os.profile_snapshot(),
-        IndexSet::just(batman.entity_address()),
-        interactors.clone(),
-    )
-    .await
-    .unwrap();
-    let outcome = outcome.per_factor.get(&bdfs.factor_source_id()).unwrap();
-    assert_eq!(outcome.to_use_directly.len(), 1);
 }
 
 #[actix_rt::test]
