+++ src/factor_instances_provider/provider/mod.rs
@@ -2,6 +2,7 @@ mod factor_instances_cache;
 mod factor_instances_provider;
 mod keyed_instances;
 mod outcome;
+mod provider_adopters;
 
 #[cfg(test)]
 mod factor_instances_provider_unit_tests;
@@ -12,3 +13,4 @@ pub use factor_instances_cache::*;
 pub use factor_instances_provider::*;
 pub use keyed_instances::*;
 pub use outcome::*;
+pub use provider_adopters::*;
